feat(admin): 实现管理员服务
- 用户列表、封禁/解封、等级调整、用户统计
- 完整单元测试
- 修复 MemoryUserStore 公钥哈希处理不一致问题
- 将 UserStats 从 election.go 移到 models.go

+++ internal/storage/badger_adapter.go
@@ -136,7 +136,7 @@ func (s *BadgerUserStore) Update(ctx context.Context, user *model.User) (*model.
 	return user, nil
 }
 
-func (s *BadgerUserStore) List(ctx context.Context, filter UserFilter) ([]*model.User, int, error) {
+func (s *BadgerUserStore) List(ctx context.Context, filter UserFilter) ([]*model.User, int64, error) {
 	users, err := s.store.ListUsers(filter.Offset, filter.Limit)
 	if err != nil {
 		return nil, 0, err
@@ -154,7 +154,7 @@ func (s *BadgerUserStore) List(ctx context.Context, filter UserFilter) ([]*model
 		filtered = append(filtered, u)
 	}
 
-	return filtered, len(filtered), nil
+	return filtered, int64(len(filtered)), nil
 }
 
 // BadgerRatingStore 适配 kv.RatingStore 到 storage.RatingStore 接口
